tools/requirements.txt
moviepy>=1.0.3
numpy>=1.21.0
imageio>=2.9.0
imageio-ffmpeg>=0.4.7
Pillow>=9.0.0
